move location of 2FA alert component
